fix build/tests?
