+++ packages/react-instantsearch-ssr-nextjs/__tests__/module/is-cjs-module.cjs
@@ -2,6 +2,7 @@
 
 const assert = require('assert');
 
+require('next');
 const ReactInstantSearchSSRNext = require('react-instantsearch-ssr-nextjs');
 
 assert.ok(ReactInstantSearchSSRNext);
