+++ packages/react-instantsearch-ssr-nextjs/__tests__/module/is-es-module.mjs
@@ -1,6 +1,7 @@
 /* eslint-disable no-console */
 import assert from 'assert';
 
+import 'next';
 import * as ReactInstantSearchSSRNext from 'react-instantsearch-ssr-nextjs';
 
 assert.ok(ReactInstantSearchSSRNext);
