+++ packages/react-instantsearch-ssr-nextjs/src/InitializePromise.tsx
@@ -1,0 +1,69 @@
+import { getInitialResults } from 'instantsearch.js/es/lib/server';
+import { walkIndex } from 'instantsearch.js/es/lib/utils';
+import { ServerInsertedHTMLContext } from 'next/navigation';
+import React, { useContext } from 'react';
+import {
+  useInstantSearchContext,
+  useRSCContext,
+  wrapPromiseWithState,
+} from 'react-instantsearch-core';
+
+export function InitializePromise() {
+  const search = useInstantSearchContext();
+  const waitForResultsRef = useRSCContext();
+  const insertHTML =
+    useContext(ServerInsertedHTMLContext) ||
+    (() => {
+      throw new Error('Missing ServerInsertedHTMLContext');
+    });
+
+  const waitForResults = () =>
+    new Promise<void>((resolve) => {
+      search.mainHelper!.derivedHelpers[0].on('result', () => {
+        resolve();
+      });
+    });
+
+  const injectInitialResults = () => {
+    const results = getInitialResults(search.mainIndex);
+    insertHTML(() => (
+      <script
+        dangerouslySetInnerHTML={{
+          __html: `(window[Symbol.for("InstantSearchInitialResults")] ??= []).push(${JSON.stringify(
+            results
+          )})`,
+        }}
+      />
+    ));
+  };
+
+  if (waitForResultsRef?.current === null) {
+    waitForResultsRef.current = wrapPromiseWithState(
+      waitForResults()
+        .then(() => {
+          let shouldRefetch = false;
+          walkIndex(search.mainIndex, (index) => {
+            shouldRefetch = index
+              .getWidgets()
+              .some((widget) => widget.$$type === 'ais.dynamicWidgets');
+          });
+
+          if (shouldRefetch) {
+            waitForResultsRef.current = wrapPromiseWithState(
+              waitForResults().then(injectInitialResults)
+            );
+          }
+
+          return shouldRefetch;
+        })
+        .then((shouldRefetch) => {
+          if (shouldRefetch) {
+            return;
+          }
+          injectInitialResults();
+        })
+    );
+  }
+
+  return null;
+}
