+++ packages/react-instantsearch-ssr-nextjs/src/TriggerSearch.ts
@@ -1,0 +1,15 @@
+import {
+  useInstantSearchContext,
+  useRSCContext,
+} from 'react-instantsearch-core';
+
+export function TriggerSearch() {
+  const instantsearch = useInstantSearchContext();
+  const waitForResultsRef = useRSCContext();
+
+  if (waitForResultsRef?.current?.status === 'pending') {
+    instantsearch.mainHelper?.searchOnlyWithDerivedHelpers();
+  }
+
+  return null;
+}
